Set aa & ad to 1000, AS to -4. Looks good with EM0
Note that setting aa and ad to 1000 IS different than setting them to 100. This is a confusing part of the documentation on the LSS communication protocol

+++ mcu/Inc/LSS/lss.h
@@ -163,7 +163,7 @@ void lss_set_hs(lss_t* hlss, int8_t holding_stiffness);
  * @param angular_accel Angular acceleration in increments of 10 deg/s/s in
  *        range [1,100]
  */
-void lss_set_aa(lss_t* hlss, uint8_t angular_accel);
+void lss_set_aa(lss_t* hlss, uint16_t angular_accel);
 
 /**
  * @brief Varies the servo's angular deceleration
@@ -171,6 +171,6 @@ void lss_set_aa(lss_t* hlss, uint8_t angular_accel);
  * @param angular_decel Angular deceleration in increments of 10 deg/s/s in
  *        range [1,100]
  */
-void lss_set_ad(lss_t* hlss, uint8_t angular_decel);
+void lss_set_ad(lss_t* hlss, uint16_t angular_decel);
 
 #endif /* LSS_H */
